% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controls.R
\name{get_control}
\alias{get_control}
\title{Print a Summary of a Maximized Monte Carlo Object}
\usage{
get_control(method, control = list())
}
\arguments{
\item{method}{A character string. Type of algorithm to be
used for global optimization. The four available methods
are simulated annealing ("GenSA"), particle swarm ("pso"),
genetic algorithm ("GA"), and grid search ("gridSearc")
Default is "GenSA",}

\item{control}{A list. Arguments to be used to control the
behavior of the algorithm chosen in \code{method}. See
details.}
}
\value{
A list. Arguments to be used to control the behavior
of the algorithm chosen in \code{method}.
}
\description{
This function
}
\keyword{internal}

