% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmc.R
\docType{package}
\name{MaxMC-package}
\alias{MaxMC-package}
\title{Maximized Monte Carlo}
\description{
Functions that implement the Maximized Monte Carlo technique based on
Dufour, J.-M. (2006), Monte Carlo Tests with nuisance parameters:
A general approach to finite sample inference and nonstandard asymptotics in econometrics.
\emph{Journal of Econometrics}, \bold{133(2)}, 443-447.

The main functions of \pkg{MaxMC} are \code{\link{mmc}} and \code{\link{mc}}.
}
\references{
Dufour, J.-M. (2006), Monte Carlo Tests with nuisance parameters:
A general approach to finite sample inference and nonstandard asymptotics in econometrics.
\emph{Journal of Econometrics}, \bold{133(2)}, 443-447.

Dufour, J.-M. and Khalaf L. (2003), Monte Carlo Test Methods in Econometrics.
in Badi H. Baltagi, ed., \emph{A Companion to Theoretical Econometrics}, Blackwell Publishing Ltd, 494-519.
}
\author{
Julien Neves, jmn252@cornell.edu (Maintainer)

Jean-Marie Dufour, jean-marie.dufour@mcgill.ca
}
