% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor.R
\name{monitor_mmc}
\alias{monitor_mmc}
\title{Format the optimization method controls}
\usage{
monitor_mmc(ite, pval, monitor)
}
\arguments{
\item{ite}{A}

\item{pval}{A}

\item{monitor}{A logical. By default, for interactive
sessions, the function gaMonitor or gaMonitor2, depending
on whether or not is an RStudio session, prints the
average and best fitness values at each iteration. If set
to plot these information are plotted on a graphical
device. Other functions can be written by the user and
supplied as argument. In non interactive sessions, by
default monitor = FALSE so any output is suppressed.}
}
\value{
A list. Arguments to be used to control the behavior
of the algorithm chosen in \code{method}.
}
\description{
This function provides a way to merge the user specified controls for the
optimization methods with their respective default controls.
}
\keyword{internal}

