% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.mmc}
\alias{plot.mmc}
\title{Plot a \code{mmc} Object}
\usage{
\method{plot}{mmc}(x, ...)
}
\arguments{
\item{x}{An object of class \code{mmc}.}

\item{...}{Arguments to be passed to methods, such as \link{graphical parameters} (see \code{\link{par}}).}
}
\value{
The \code{mmc} object is returned invisibly.
}
\description{
The \code{plot()} method for objects of the
class \code{mmc} gives a plot of the best and current p-value found during the iterations of \code{mmc}.
}
