% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.mc}
\alias{print.mc}
\title{Print a Summary of a Monte Carlo Object}
\usage{
\method{print}{mc}(x, digits = getOption("digits"), ...)
}
\arguments{
\item{x}{an object used to select a method.}

\item{digits}{minimal number of \emph{significant} digits, see
    \code{\link{print.default}}.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
This is a method for the function print() for objects of the
class "mmc".
}

