% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvalue.R
\name{pvalue}
\alias{pvalue}
\title{P-value function}
\usage{
pvalue(S0, S, type = c("geq", "leq", "absolute", "two-tailed"))
}
\arguments{
\item{S0}{An atomic vector. Value of the test statistic
applied to the data.}

\item{S}{An atomic vector. Replication of the test statistic.
S must have length greater than one, with no missing values.}

\item{type}{A character string. Specifies the type of test
the p-value function produce. The possible values are
\code{geq}, \code{leq}, \code{absolute} and \code{two-tailed}.
 Default is \code{geq}.}
}
\value{
The p-value of S0 given the vector of replication S.
}
\description{
Computes the p-value of the statistic by computing its rank
compared to its simulated values.
}
\details{
We allow for three-type of p-value: \code{leq}, \code{geq},
\code{absolute} and \code{two-tailed}. For one-tailed test,
\code{leq} returns the proportion of simulated values smaller
 than the statistic while \code{geq} return the proportion of
 simulated values greater than the statistic. For two-tailed
 test, if the statistic is symmetric, one can use the
 absolute value of the statistic and its simulated value to
 retrieve a two-tailed test. If the statistic is not
 symmetric,
 one can specify the p-value type as \code{two-tailed} to
 obtained the p-value computed as the twice the minimum of
 \code{leq} and \code{geq}.

Ties in the ranking are broken according to a uniform
distribution.
}

