% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{return_mc}
\alias{return_mc}
\title{S3 class mc.}
\usage{
return_mc(S0, y, statistic, dgp, N, type, call, seed, pval)
}
\arguments{
\item{S0}{An atomic vector. Value of the test statistic
applied to the data.}

\item{y}{A vector or data frame.}

\item{statistic}{A function or a character string. Specifies
how the statistic is computed. The function inputs the
\code{y}, and outputs a scalar numeric.}

\item{dgp}{A function. The function inputs the first argument
\code{y}, and outputs a simulated \code{y} (i.e. an object
of the same type as \code{y}). It is analogous to the data
generating process under the null. Default value is
sample(y, replace = TRUE), the bootstrap resampling of
 \code{y}.}

\item{N}{An atomic vector. Number of replication of test
statistic.}

\item{type}{A character string. Specifies the type of test
the p-value function produce. The possible values are
\code{geq}, \code{leq}, \code{absolute} and \code{two-tailed}.
 Default is \code{geq}.}

\item{call}{Original call to \code{mmc}}

\item{seed}{value of \code{.Random.seed} at the start of
\code{mmc} call}

\item{pval}{Monte Carlo p-value of \code{statistic}}
}
\description{
S3 class mc.
}
\keyword{internal}

