% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{return_mmc}
\alias{return_mmc}
\title{S3 class mmc.}
\usage{
return_mmc(S0, y, statistic, dgp, est, lower, upper, N, type, method, alpha,
  control, call, seed, lmc, opt_result)
}
\arguments{
\item{S0}{An atomic vector. Value of the test statistic
applied to the data.}

\item{y}{A vector or data frame.}

\item{statistic}{A function or a character string. Specifies
how the statistic is computed. The function inputs the
\code{y}, and outputs a scalar numeric.}

\item{dgp}{A function. The function inputs the first argument
\code{y} and a vector of nuissance paramaters \code{v}, and
outputs a simulated \code{y} (i.e. an object of the same type
as \code{y}). It is analogous to the data generating process
under the null. The function can be specified without
\code{v}. Default value is sample(y, replace = TRUE), the
 bootstrap resampling of \code{y}.}

\item{est}{A vector with length of v. Default is NULL, in
which case, default values will be generated automatically.}

\item{lower}{A vector with length of v. Lower bounds for
nuisance parameters under the null}

\item{upper}{A vector with length of v. Upper bounds for
nuisance parameters under the null}

\item{N}{An atomic vector. Number of replication of test
statistic.}

\item{type}{A character string. Specifies the type of test
the p-value function produce. The possible values are
\code{geq}, \code{leq}, \code{absolute} and \code{two-tailed}.
 Default is \code{geq}.}

\item{method}{A character string. Type of algorithm to be
used for global optimization. The four available methods
are simulated annealing ("GenSA"), particle swarm ("pso"),
genetic algorithm ("GA"), and grid search ("gridSearc")
Default is "GenSA",}

\item{alpha}{A atomic vector. Default is NULL.}

\item{control}{A list. Arguments to be used to control the
behavior of the algorithm chosen in \code{method}. See
details.}

\item{call}{Original call to \code{mmc}}

\item{seed}{value of \code{.Random.seed} at the start of
\code{mmc} call}

\item{lmc}{if \code{par} is specified, it returns an object
of class "\code{mc}" corresponding the Local Monte Carlo test}

\item{opt_result}{an object returning the optimization results}
}
\description{
S3 class mmc.
}
\keyword{internal}

