% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulation_mmc}
\alias{simulation_mmc}
\title{Maximized Monte Carlo Simulation}
\usage{
simulation_mmc(y, statistic, dgp = function(y, v) sample(y, replace = TRUE),
  v, N = 99, ...)
}
\arguments{
\item{y}{A vector or data frame.}

\item{statistic}{A function or a character string. Specifies
how the statistic is computed. The function inputs the
\code{y}, and outputs a scalar numeric.}

\item{dgp}{A function. The function inputs the first argument
\code{y} and a vector of nuisance paramaters \code{v}, and
outputs a simulated \code{y} (i.e. an object of the same type
as \code{y}). It is important to respect the order of the arguments.
It should represent the data generating process under the null.
The function can be specified without any nuisance parameters
(i.e. without \code{v}). Default value is sample(y, replace = TRUE),
the bootstrap resampling of \code{y}.}

\item{v}{A vector paramaters. The vector \code{v} is use to
specify the \code{dgp}. Note that if \code{dgp} is a
function  of only \code{y} then we do not need to specify
\code{v}. Default value is NULL.}

\item{N}{An atomic vector. Number of replications of the test
statistic.}

\item{...}{Other named arguments for statistic which are
passed unchanged each time it is called}
}
\value{
The vector of replication of test statistic.
}
\description{
Generates N Monte Carlo replicates of a statistic for given nuisance parameter value.
}
\keyword{internal}
