% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor.R
\name{monitor_mmc}
\alias{monitor_mmc}
\title{Format the optimization method controls}
\usage{
monitor_mmc(object, alpha = NULL, monitor = TRUE)
}
\arguments{
\item{object}{A data.frame}

\item{alpha}{An atomic vector. If \code{mmc} finds a
p-value over \code{alpha}, then the algorithm will stop. This is particularly
useful if we are only looking at testing a hypothesis at a particular level.
Default is NULL.}

\item{monitor}{A logical. If set to TRUE, then the p-values at every
iteration and the cumulative maximum p-value are plotted on a graphical device.
Default is FALSE.}
}
\value{
A list. Arguments to be used to control the behavior
of the algorithm chosen in \code{method}.
}
\description{
This function provides a way to merge the user specified controls for the
optimization methods with their respective default controls.
}
\keyword{internal}
