% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.mmc}
\alias{plot.mmc}
\title{Print a Summary of a \code{mmc} Object}
\usage{
\method{plot}{mmc}(x, ...)
}
\arguments{
\item{x}{An object of class "mmc"}

\item{...}{Arguments to be passed to methods, such as \link{graphical parameters} (see \code{\link{par}}).}
}
\value{
The \code{mmc} object is returned invisibly.
}
\description{
This is a method for the function \code{print()} for objects of the
class \code{mmc}.
}
