% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.mc}
\alias{print.mc}
\title{Print a Summary of a \code{mc} Object}
\usage{
\method{print}{mc}(x, digits = getOption("digits"), ...)
}
\arguments{
\item{x}{an object used to select a method.}

\item{digits}{minimal number of \emph{significant} digits, see
    \code{\link{print.default}}.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
The \code{mc} object is returned invisibly.
}
\description{
This is a method for the function \code{print()} for objects of the
class \code{mmc}.
}

