% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.mmc}
\alias{print.mmc}
\title{Print a Summary of a \code{mmc} Object}
\usage{
\method{print}{mmc}(x, digits = getOption("digits"), ...)
}
\arguments{
\item{x}{an object used to select a method.}

\item{digits}{minimal number of \emph{significant} digits, see
    \code{\link{print.default}}.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
The \code{mmc} object is returned invisibly.
}
\description{
This is a method for the function \code{print()} for objects of the
class \code{mmc}.
}
\examples{
## Example
library(fUnitRoots)
# Set seed
set.seed(123)

# Generate an AR(2) process with phi = (-1.5,0.5), and n = 25
y <- filter(rnorm(25), c(-1.5, 0.5), method = "recursive")

# Set bounds for the nuisance parameter v
lower <- -1
upper <- 1

# Set the function to generate an AR(2) integrated process
dgp <- function(y, v) {
    ran.y <- filter(rnorm(length(y)), c(1-v,v), method = "recursive")
}

# Set the Augmented-Dicky Fuller statistic
statistic <- function(y){
    out <- suppressWarnings(adfTest(y, lags = 2, type = "nc"))
    return(out@test$statistic)
}

# Apply the mmc procedure
est <- mmc(y, statistic = statistic , dgp = dgp, lower = lower,
           upper = upper, N = 99, type = "leq", method = "GenSA",
           control = list(max.time = 2))

# Print result of object of class 'mmc'
print(est)
}
